+++ checker/src/main/java/org/checkerframework/checker/lock/LockVisitor.java
@@ -37,7 +37,6 @@ import org.checkerframework.framework.type.AnnotatedTypeMirror;
 import org.checkerframework.framework.type.AnnotatedTypeMirror.AnnotatedDeclaredType;
 import org.checkerframework.framework.type.AnnotatedTypeMirror.AnnotatedExecutableType;
 import org.checkerframework.framework.type.QualifierHierarchy;
-import org.checkerframework.framework.util.AnnotatedTypes;
 import org.checkerframework.framework.util.JavaExpressionParseUtil.JavaExpressionParseException;
 import org.checkerframework.framework.util.StringToJavaExpression;
 import org.checkerframework.framework.util.dependenttypes.DependentTypesError;
@@ -640,9 +639,7 @@ public class LockVisitor extends BaseTypeVisitor<LockAnnotatedTypeFactory> {
         ParameterizedExecutableType mType = atypeFactory.methodFromUse(methodInvocationTree);
         AnnotatedExecutableType invokedMethod = mType.executableType;
 
-        List<AnnotatedTypeMirror> paramTypes =
-                AnnotatedTypes.adaptParameters(
-                        atypeFactory, invokedMethod, methodInvocationTree.getArguments());
+        List<AnnotatedTypeMirror> paramTypes = invokedMethod.getParameterTypes();
 
         // Index on @GuardSatisfied at each location. -1 when no @GuardSatisfied annotation was
         // present.