+++ framework/src/main/java/org/checkerframework/common/basetype/BaseTypeVisitor.java
@@ -1775,9 +1775,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
                     tree.getTypeArguments(),
                     methodName,
                     invokedMethod.getTypeVariables());
-            List<AnnotatedTypeMirror> params =
-                    AnnotatedTypes.adaptParameters(
-                            atypeFactory, invokedMethod, tree.getArguments());
+            List<AnnotatedTypeMirror> params = invokedMethod.getParameterTypes();
             checkArguments(params, tree.getArguments(), methodName, method.getParameters());
             checkVarargs(invokedMethod, tree);
 
@@ -1896,11 +1894,8 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
             return;
         }
 
-        List<AnnotatedTypeMirror> formals = invokedMethod.getParameterTypes();
-        int numFormals = formals.size();
-        int lastArgIndex = numFormals - 1;
         // This is the varags type, an array.
-        AnnotatedArrayType lastParamAnnotatedType = (AnnotatedArrayType) formals.get(lastArgIndex);
+        AnnotatedArrayType lastParamAnnotatedType = invokedMethod.getVarargType();
 
         AnnotatedTypeMirror wrappedVarargsType = atypeFactory.getAnnotatedTypeVarargsArray(tree);
 
@@ -2097,8 +2092,7 @@ public class BaseTypeVisitor<Factory extends GenericAnnotatedTypeFactory<?, ?, ?
         List<AnnotatedTypeMirror> typeargs = fromUse.typeArgs;
 
         List<? extends ExpressionTree> passedArguments = tree.getArguments();
-        List<AnnotatedTypeMirror> params =
-                AnnotatedTypes.adaptParameters(atypeFactory, constructorType, passedArguments);
+        List<AnnotatedTypeMirror> params = constructorType.getParameterTypes();
 
         ExecutableElement constructor = constructorType.getElement();
         CharSequence constructorName = ElementUtils.getSimpleNameOrDescription(constructor);