+++ framework/src/main/java/org/checkerframework/framework/type/treeannotator/PropagationTreeAnnotator.java
@@ -142,8 +142,7 @@ public class PropagationTreeAnnotator extends TreeAnnotator {
 
                 // If the method accepts varargs, we handle the vararg parameter after this
                 // for-loop.
-                int loopBound = isVarargs ? parametersCount - 1 : parametersCount;
-                for (int i = 0; i < loopBound; i++) {
+                for (int i = 0; i < parametersCount; i++) {
                     @SuppressWarnings("interning") // Tree must be exactly the same.
                     boolean foundArgument = methodInvocationTree.getArguments().get(i) == tree;
                     if (foundArgument) {
@@ -160,9 +159,7 @@ public class PropagationTreeAnnotator extends TreeAnnotator {
                     // any explicit arguments for the varargs (i.e., an empty array).
                     // So we don't try getting the argument from methodInvocationTree
                     // to avoid out-of-bound exception.
-                    // TODO: argument could be the varargs array vs one of the elements
-                    // in that array?
-                    contextType = m.getParameterTypes().get(parametersCount - 1);
+                    contextType = m.getVarargType();
                 }
             }
         }