+++ framework/src/test/java/org/checkerframework/framework/test/junit/H1H2CheckerTest.java
@@ -7,7 +7,6 @@ import org.junit.runners.Parameterized.Parameters;
 import java.io.File;
 import java.util.List;
 
-/** */
 public class H1H2CheckerTest extends CheckerFrameworkPerDirectoryTest {
 
     /**
@@ -18,7 +17,8 @@ public class H1H2CheckerTest extends CheckerFrameworkPerDirectoryTest {
                 testFiles,
                 H1H2Checker.class,
                 "h1h2checker",
-                "-Astubs=tests/h1h2checker/h1h2checker.astub");
+                "-Astubs=tests/h1h2checker/h1h2checker.astub",
+                "-AcheckEnclosingExpr");
     }
 
     @Parameters