+++ framework/src/test/java/org/checkerframework/framework/test/junit/FrameworkTest.java
@@ -1,17 +1,18 @@
 package org.checkerframework.framework.test.junit;
 
-import org.checkerframework.framework.test.CheckerFrameworkPerFileTest;
+import org.checkerframework.framework.test.CheckerFrameworkPerDirectoryTest;
 import org.checkerframework.framework.testchecker.util.EvenOddChecker;
 import org.junit.runners.Parameterized.Parameters;
 
 import java.io.File;
+import java.util.List;
 
 /** JUnit tests for the Checker Framework, using the {@link EvenOddChecker}. */
-// See framework/tests/framework/README for why this is a CFPerFileTest.
-public class FrameworkTest extends CheckerFrameworkPerFileTest {
+// See FrameworkJavacErrorsTest for tests that can contain Java errors.
+public class FrameworkTest extends CheckerFrameworkPerDirectoryTest {
 
-    public FrameworkTest(File testFile) {
-        super(testFile, EvenOddChecker.class, "framework");
+    public FrameworkTest(List<File> testFiles) {
+        super(testFiles, EvenOddChecker.class, "framework");
     }
 
     @Parameters