+++ javacutil/src/main/java/org/checkerframework/javacutil/TreeUtils.java
@@ -43,6 +43,7 @@ import com.sun.tools.javac.tree.JCTree.JCAnnotation;
 import com.sun.tools.javac.tree.JCTree.JCBinary;
 import com.sun.tools.javac.tree.JCTree.JCExpression;
 import com.sun.tools.javac.tree.JCTree.JCExpressionStatement;
+import com.sun.tools.javac.tree.JCTree.JCFieldAccess;
 import com.sun.tools.javac.tree.JCTree.JCLambda;
 import com.sun.tools.javac.tree.JCTree.JCLambda.ParameterKind;
 import com.sun.tools.javac.tree.JCTree.JCLiteral;
@@ -115,6 +116,9 @@ public final class TreeUtils {
     /** Whether we are running on at least Java 16. */
     private static final boolean atLeastJava16;
 
+    /** Whether we are running on at least Java 21. */
+    private static final boolean atLeastJava21;
+
     /** The CaseTree.getExpression method for Java up to 11; null otherwise. */
     private static final @Nullable Method CASETREE_GETEXPRESSION;
 
@@ -143,6 +147,12 @@ public final class TreeUtils {
 
     /** The BindingPatternTree.getVariable method for Java 16 and higher; null otherwise. */
     private static final @Nullable Method BINDINGPATTERNTREE_GETVARIABLE;
+
+    /**
+     * The {@code TreeMaker.Select(JCExpression, Symbol)} method. Return type changes for JDK21+.
+     * Only needs to be used while the code is compiled with JDK below 21.
+     */
+    private static final @Nullable Method TREEMAKER_SELECT;
 
     /** The value of Flags.RECORD which does not exist in Java 9 or 11. */
     private static final long Flags_RECORD = 2305843009213693952L;
@@ -182,6 +192,14 @@ public final class TreeUtils {
             java16 = null;
         }
         atLeastJava16 = java16 != null && latestSource.ordinal() >= java16.ordinal();
+
+        SourceVersion java21;
+        try {
+            java21 = SourceVersion.valueOf("RELEASE_21");
+        } catch (IllegalArgumentException e) {
+            java21 = null;
+        }
+        atLeastJava21 = java21 != null && latestSource.ordinal() >= java21.ordinal();
 
         try {
             // TODO: profile and see whether doing all these here has a performance impact.
@@ -241,6 +259,12 @@ public final class TreeUtils {
             } else {
                 INSTANCEOFTREE_GETPATTERN = null;
                 BINDINGPATTERNTREE_GETVARIABLE = null;
+            }
+            if (atLeastJava21) {
+                TREEMAKER_SELECT =
+                        TreeMaker.class.getMethod("Select", JCExpression.class, Symbol.class);
+            } else {
+                TREEMAKER_SELECT = null;
             }
         } catch (ClassNotFoundException | NoSuchMethodException e) {
             Error err = new AssertionError("Unexpected error in TreeUtils static initializer");
@@ -2544,5 +2568,51 @@ public final class TreeUtils {
      */
     public static boolean isBinaryComparison(BinaryTree tree) {
         return BINARY_COMPARISON_TREE_KINDS.contains(tree.getKind());
+    }
+
+    /**
+     * Returns the result of {@code treeMaker.Select(base, sym)}.
+     *
+     * @param treeMaker the TreeMaker to use
+     * @param base the expression for the select
+     * @param sym the symbol to select
+     * @return the JCFieldAccess tree to select sym in base
+     */
+    public static JCFieldAccess Select(TreeMaker treeMaker, Tree base, Symbol sym) {
+        if (atLeastJava21) {
+            try {
+                assert TREEMAKER_SELECT != null : "@AssumeAssertion(nullness): initialization";
+                JCFieldAccess jfa = (JCFieldAccess) TREEMAKER_SELECT.invoke(treeMaker, base, sym);
+                if (jfa != null) {
+                    return jfa;
+                } else {
+                    throw new BugInCF(
+                            "TreeUtils.Select: TreeMaker.Select returned null for tree: %s", base);
+                }
+            } catch (InvocationTargetException | IllegalAccessException e) {
+                throw new BugInCF("TreeUtils.Select: reflection failed for tree: %s", base, e);
+            }
+        } else {
+            @SuppressWarnings("cast") // Redundant on JDK 21+
+            JCFieldAccess jfa = (JCFieldAccess) treeMaker.Select((JCExpression) base, sym);
+            return jfa;
+        }
+    }
+
+    /**
+     * Returns the result of {@code treeMaker.Select(base, name)}.
+     *
+     * @param treeMaker the TreeMaker to use
+     * @param base the expression for the select
+     * @param name the name to select
+     * @return the JCFieldAccess tree to select sym in base
+     */
+    public static JCFieldAccess Select(
+            TreeMaker treeMaker, Tree base, com.sun.tools.javac.util.Name name) {
+        /*
+         * There's no need for reflection here. The only reason we even declare this method is so that
+         * callers don't have to remember which overload we provide a wrapper around.
+         */
+        return treeMaker.Select((JCExpression) base, name);
     }
 }