+++ javacutil/src/main/java/org/checkerframework/javacutil/trees/TreeBuilder.java
@@ -125,9 +125,7 @@ public class TreeBuilder {
                         com.sun.tools.javac.util.List.nil(),
                         methodClass);
 
-        JCTree.JCFieldAccess iteratorAccess =
-                (JCTree.JCFieldAccess)
-                        maker.Select((JCTree.JCExpression) iterableExpr, iteratorMethod);
+        JCTree.JCFieldAccess iteratorAccess = TreeUtils.Select(maker, iterableExpr, iteratorMethod);
         iteratorAccess.setType(updatedMethodType);
 
         return iteratorAccess;
@@ -156,11 +154,10 @@ public class TreeBuilder {
             }
         }
 
-        assert hasNextMethod != null : "no hasNext method declared for expression type";
+        assert hasNextMethod != null
+                : "@AssumeAssertion(nullness): no hasNext method declared for expression type";
 
-        JCTree.JCFieldAccess hasNextAccess =
-                (JCTree.JCFieldAccess)
-                        maker.Select((JCTree.JCExpression) iteratorExpr, hasNextMethod);
+        JCTree.JCFieldAccess hasNextAccess = TreeUtils.Select(maker, iteratorExpr, hasNextMethod);
         hasNextAccess.setType(hasNextMethod.asType());
 
         return hasNextAccess;
@@ -210,8 +207,7 @@ public class TreeBuilder {
                         com.sun.tools.javac.util.List.nil(),
                         methodClass);
 
-        JCTree.JCFieldAccess nextAccess =
-                (JCTree.JCFieldAccess) maker.Select((JCTree.JCExpression) iteratorExpr, nextMethod);
+        JCTree.JCFieldAccess nextAccess = TreeUtils.Select(maker, iteratorExpr, nextMethod);
         nextAccess.setType(updatedMethodType);
 
         return nextAccess;
@@ -225,8 +221,7 @@ public class TreeBuilder {
      */
     public MemberSelectTree buildArrayLengthAccess(ExpressionTree expression) {
 
-        return (JCTree.JCFieldAccess)
-                maker.Select((JCTree.JCExpression) expression, symtab.lengthVar);
+        return TreeUtils.Select(maker, expression, symtab.lengthVar);
     }
 
     /**
@@ -406,8 +401,7 @@ public class TreeBuilder {
 
         Type.MethodType methodType = (Type.MethodType) valueOfMethod.asType();
 
-        JCTree.JCFieldAccess valueOfAccess =
-                (JCTree.JCFieldAccess) maker.Select((JCTree.JCExpression) expr, valueOfMethod);
+        JCTree.JCFieldAccess valueOfAccess = TreeUtils.Select(maker, expr, valueOfMethod);
         valueOfAccess.setType(methodType);
 
         return valueOfAccess;
@@ -467,8 +461,7 @@ public class TreeBuilder {
 
         Type.MethodType methodType = (Type.MethodType) primValueMethod.asType();
 
-        JCTree.JCFieldAccess primValueAccess =
-                (JCTree.JCFieldAccess) maker.Select((JCTree.JCExpression) expr, primValueMethod);
+        JCTree.JCFieldAccess primValueAccess = TreeUtils.Select(maker, expr, primValueMethod);
         primValueAccess.setType(methodType);
 
         return primValueAccess;