+++ docs/CHANGELOG.md
@@ -1,3 +1,13 @@
+Version 3.32.0-eisop2 (March ?, 2023)
+-------------------------------------
+
+**User-visible changes:**
+
+**Implementation details:**
+
+**Closed issues:**
+
+
 Version 3.32.0-eisop1 (March 9, 2023)
 -------------------------------------